Extracted Grid<Book> listener & its method to BookGridListener
